backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.2
sqlalchemy==2.0.27
requests==2.31.0
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
alembic==1.13.1
psycopg2-binary==2.9.9  # For PostgreSQL support (optional)
bcrypt==4.1.2
pydantic[email]==2.6.2
pyjwt==2.1.0
gunicorn==21.2.0
tenacity==8.2.3
aiohttp==3.9.3
pytest==7.4.3
httpx==0.26.0
sentry-sdk==1.39.1
redis==5.0.2
